requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
ExifRead==3.0.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
importlib==1.0.4
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pytest==8.3.3
python-engineio==4.9.1
python-socketio==5.11.4
simple-websocket==1.1.0
Werkzeug==3.0.4
wsproto==1.2.0
